+++ package.json
@@ -74,7 +74,7 @@
     "tsc-alias": "^1.8.16",
     "typescript": "^5.9.3",
     "vite": "^7.3.5",
-    "vite-tsconfig-paths": "^6.0.3",
+    "vite-tsconfig-paths": "^6.1.1",
     "vitest": "^3.2.0"
   },
   "peerDependencies": {
