+++ src/components/MIStepper/__test__/MIStep.test.tsx
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { render } from '../../../test-utils';
 
 import MIStep from '../MIStep';
