+++ src/components/MIStepper/__test__/MIStepper.test.tsx
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { render } from '../../../test-utils';
 
 import { createMatchMedia } from '../../../test-utils';
 import MIStepper from '../MIStepper';
