+++ tsconfig.prod.json
@@ -15,7 +15,7 @@
     "src/**/*.spec.ts",
     "src/**/*.spec.tsx",
     "src/**/__tests__/**/*",
-    "src/test-utils.ts",
+    "src/test-utils.tsx",
     "vite.config.mts",
     "vitest.setup.ts"
   ]
